[ST-6] TodoManager 삭제 기능 구현 - TodoManager.delete(Todo) 메소드 구현 - Testcase 분리
