+++ src/main/java/net/slipp/todo_list/TodoRepository.java
@@ -9,5 +9,6 @@ public class TodoRepository {
 
     public Todo store(Todo todo) throws IllegalArgumentException, RepositoryFailedException { return null; }
 
+    public Todo store(Todo todo, boolean should_delete) throws IllegalArgumentException { return null; }
 
 }
